Feat/gitignore sectioned sorter (#17)
* refactor: move file resolver into shared.

* refactor: move file resolver into shared.

* refactor: set up locations and automation for new hook.

* feat: add argparse for sort_gitignore

* refactor: rename sort_gitignore to sort_file_contents.

The convention of passing .gitignore as the file argument to a git hook results in some file containing the word gitignore in their names getting sorted as well. This changes avoids that while also reflecting the fact that the hook has utility beyond simply gitignore.

* refactor: rename sort_gitignore to sort_file_contents.

The convention of passing .gitignore as the file argument to a git hook results in some file containing the word gitignore in their names getting sorted as well. This changes avoids that while also reflecting the fact that the hook has utility beyond simply gitignore.

* feat: file reading

* feat: separate a list of strings into sections.

* fix: remove blank lines in early return.

* feat: separate leading comment from section.

* feat: line sorting.

* feat: functioning executable

* docs: update WIP docstring.

* feat: ignore comment characters when sorting.